+++ src/spin_out.h
@@ -8,6 +8,7 @@
 
 #include "MC_structure.h"
 
+int WriteSpin(Supercell & supercell, std::string spin_structure_file_prefix);
 int WriteSpin(Supercell & supercell, std::string spin_structure_file_prefix, double T);
 
 #endif